chore(tests): scope numpy rng per function in tests, rather than per session, to ensure reproducibility of test pipeline

+++ vpop_calibration/pynlme/data.py
@@ -68,7 +68,7 @@ class ObsData:
                 torch.ones_like(
                     index_values_p.output_name.index_values,
                     device=device,
-                    dtype=torch.float64,
+                    dtype=default_dtype,
                 ),
             )
